Move pandas distributional expectations to new submodule
