+++ tests/__init__.py
@@ -2,6 +2,6 @@ from .test_great_expectations import *
 from .test_util import *
 from .test_dataset import *
 from .test_pandas_dataset import *
-from .test_pandas_dataset_distributional_expectations import *
+from tests.pandas.test_pandas_dataset_distributional_expectations import *
 from .test_expectation_decorators import *
 from .test_cli import *
